Moved the data slice helper classes to Utils/Volume project, and made separate classes for data and mask volumes.

+++ src/Application/Layer/LayerGroup.h
@@ -179,7 +179,7 @@ public:
   {
     return layer_list_;
   }
-  int get_list_size()
+  size_t get_list_size()
   {
     return layer_list_.size();
   }
